Introduce Request namespace in Query
