+++ spec/Query/Repository/Request/AllBranchStatusesQuerySpec.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace spec\DevboardLib\GitHubApi\Query\Repository;
+namespace spec\DevboardLib\GitHubApi\Query\Repository\Request;
 
 use DevboardLib\GitHub\Repo\RepoFullName;
 use DevboardLib\GitHubApi\Credentials\InstallationCredentials;
-use DevboardLib\GitHubApi\Query\Repository\AllPullRequestStatusesQuery;
+use DevboardLib\GitHubApi\Query\Repository\Request\AllBranchStatusesQuery;
 use PhpSpec\ObjectBehavior;
 
-class AllPullRequestStatusesQuerySpec extends ObjectBehavior
+class AllBranchStatusesQuerySpec extends ObjectBehavior
 {
     public function let(RepoFullName $repoFullName, InstallationCredentials $credentials)
     {
@@ -18,6 +18,6 @@ class AllPullRequestStatusesQuerySpec extends ObjectBehavior
 
     public function it_is_initializable()
     {
-        $this->shouldHaveType(AllPullRequestStatusesQuery::class);
+        $this->shouldHaveType(AllBranchStatusesQuery::class);
     }
 }
