+++ spec/Query/Repository/Request/AllBranchesQuerySpec.php
@@ -2,14 +2,15 @@
 
 declare(strict_types=1);
 
-namespace spec\DevboardLib\GitHubApi\Query\Repository;
+namespace spec\DevboardLib\GitHubApi\Query\Repository\Request;
 
 use DevboardLib\GitHub\Repo\RepoFullName;
 use DevboardLib\GitHubApi\Credentials\InstallationCredentials;
-use DevboardLib\GitHubApi\Query\Repository\AllLabelsQuery;
+use DevboardLib\GitHubApi\Query\Repository\Request\AllBranchesQuery;
+use DevboardLib\GitHubApi\Query\Repository\Request\RepositoryQuery;
 use PhpSpec\ObjectBehavior;
 
-class AllLabelsQuerySpec extends ObjectBehavior
+class AllBranchesQuerySpec extends ObjectBehavior
 {
     public function let(RepoFullName $repoFullName, InstallationCredentials $credentials)
     {
@@ -18,6 +19,7 @@ class AllLabelsQuerySpec extends ObjectBehavior
 
     public function it_is_initializable()
     {
-        $this->shouldHaveType(AllLabelsQuery::class);
+        $this->shouldHaveType(AllBranchesQuery::class);
+        $this->shouldImplement(RepositoryQuery::class);
     }
 }
