+++ spec/Query/Repository/Request/AllLabelsQuerySpec.php
@@ -2,15 +2,14 @@
 
 declare(strict_types=1);
 
-namespace spec\DevboardLib\GitHubApi\Query\Repository;
+namespace spec\DevboardLib\GitHubApi\Query\Repository\Request;
 
 use DevboardLib\GitHub\Repo\RepoFullName;
 use DevboardLib\GitHubApi\Credentials\InstallationCredentials;
-use DevboardLib\GitHubApi\Query\Repository\AllBranchesQuery;
-use DevboardLib\GitHubApi\Query\Repository\RepositoryQuery;
+use DevboardLib\GitHubApi\Query\Repository\Request\AllLabelsQuery;
 use PhpSpec\ObjectBehavior;
 
-class AllBranchesQuerySpec extends ObjectBehavior
+class AllLabelsQuerySpec extends ObjectBehavior
 {
     public function let(RepoFullName $repoFullName, InstallationCredentials $credentials)
     {
@@ -19,7 +18,6 @@ class AllBranchesQuerySpec extends ObjectBehavior
 
     public function it_is_initializable()
     {
-        $this->shouldHaveType(AllBranchesQuery::class);
-        $this->shouldImplement(RepositoryQuery::class);
+        $this->shouldHaveType(AllLabelsQuery::class);
     }
 }
