+++ spec/Query/Repository/Request/AllMilestonesQuerySpec.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace spec\DevboardLib\GitHubApi\Query\Repository;
+namespace spec\DevboardLib\GitHubApi\Query\Repository\Request;
 
 use DevboardLib\GitHub\Repo\RepoFullName;
 use DevboardLib\GitHubApi\Credentials\InstallationCredentials;
-use DevboardLib\GitHubApi\Query\Repository\AllBranchStatusesQuery;
+use DevboardLib\GitHubApi\Query\Repository\Request\AllMilestonesQuery;
 use PhpSpec\ObjectBehavior;
 
-class AllBranchStatusesQuerySpec extends ObjectBehavior
+class AllMilestonesQuerySpec extends ObjectBehavior
 {
     public function let(RepoFullName $repoFullName, InstallationCredentials $credentials)
     {
@@ -18,6 +18,6 @@ class AllBranchStatusesQuerySpec extends ObjectBehavior
 
     public function it_is_initializable()
     {
-        $this->shouldHaveType(AllBranchStatusesQuery::class);
+        $this->shouldHaveType(AllMilestonesQuery::class);
     }
 }
