+++ src/V4/Object/Repository/MilestoneObjectApi.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace DevboardLib\GitHubApi\V4\Object\Repository;
 
-use DevboardLib\GitHubApi\Query\Repository\AllMilestonesQuery;
+use DevboardLib\GitHubApi\Query\Repository\Request\AllMilestonesQuery;
 use DevboardLib\GitHubApi\V4\Object\Repository\Factory\MilestoneFactory;
 use DevboardLib\GitHubApi\V4\Object\Repository\Result\AllMilestonesResult;
 use DevboardLib\GitHubApi\V4\Raw\Repository\MilestoneApi;
