+++ src/V4/Object/Repository/StatusObjectApi.php
@@ -7,8 +7,8 @@ namespace DevboardLib\GitHubApi\V4\Object\Repository;
 use DevboardLib\Git\Branch\BranchName;
 use DevboardLib\Git\Commit\CommitSha;
 use DevboardLib\GitHub\PullRequest\PullRequestNumber;
-use DevboardLib\GitHubApi\Query\Repository\AllBranchStatusesQuery;
-use DevboardLib\GitHubApi\Query\Repository\AllPullRequestStatusesQuery;
+use DevboardLib\GitHubApi\Query\Repository\Request\AllBranchStatusesQuery;
+use DevboardLib\GitHubApi\Query\Repository\Request\AllPullRequestStatusesQuery;
 use DevboardLib\GitHubApi\V4\Object\Repository\Factory\StatusFactory;
 use DevboardLib\GitHubApi\V4\Object\Repository\Response\BranchStatusCollection;
 use DevboardLib\GitHubApi\V4\Object\Repository\Response\PullRequestStatusCollection;
