+++ src/V4/Raw/Repository/MilestoneApi.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace DevboardLib\GitHubApi\V4\Raw\Repository;
 
-use DevboardLib\GitHubApi\Query\Repository\AllMilestonesQuery;
+use DevboardLib\GitHubApi\Query\Repository\Request\AllMilestonesQuery;
 use DevboardLib\GitHubApi\V3\GitHubClientFactory;
 
 class MilestoneApi
