+++ tests/V4/TestData/TestDataRetrieverV4Test.php
@@ -9,12 +9,12 @@ use DevboardLib\GitHub\Repo\RepoFullName;
 use DevboardLib\GitHub\User\UserId;
 use DevboardLib\GitHubApi\Auth\GitHubApp\JwtTokenBuilder;
 use DevboardLib\GitHubApi\Credentials\InstallationCredentials;
-use DevboardLib\GitHubApi\Query\Repository\AllBranchesQuery;
-use DevboardLib\GitHubApi\Query\Repository\AllBranchStatusesQuery;
-use DevboardLib\GitHubApi\Query\Repository\AllLabelsQuery;
-use DevboardLib\GitHubApi\Query\Repository\AllMilestonesQuery;
-use DevboardLib\GitHubApi\Query\Repository\AllPullRequestsQuery;
-use DevboardLib\GitHubApi\Query\Repository\AllPullRequestStatusesQuery;
+use DevboardLib\GitHubApi\Query\Repository\Request\AllBranchesQuery;
+use DevboardLib\GitHubApi\Query\Repository\Request\AllBranchStatusesQuery;
+use DevboardLib\GitHubApi\Query\Repository\Request\AllLabelsQuery;
+use DevboardLib\GitHubApi\Query\Repository\Request\AllMilestonesQuery;
+use DevboardLib\GitHubApi\Query\Repository\Request\AllPullRequestsQuery;
+use DevboardLib\GitHubApi\Query\Repository\Request\AllPullRequestStatusesQuery;
 use DevboardLib\GitHubApi\V3\GitHubClientFactory;
 use DevboardLib\GitHubApi\V4\Raw\Repository\BranchApi;
 use DevboardLib\GitHubApi\V4\Raw\Repository\LabelApi;
